requirements.txt
# requirements.txt
Flask==3.0.3
pytest==8.3.2
Flask-SQLAlchemy==3.1.1
requests==2.32.3  # NEW: for HTTP calls to odds/scores providers
python-dotenv>=1.0
httpx>=0.27
